moved files from SmokeTest to Support
