+++ source/DasBlog.Tests/Support/DasBlog.Tests.Support/WaitService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading;
 
-namespace DasBlog.Tests.SmokeTest.Support
+namespace DasBlog.Tests.Support
 {
 	internal class WaitService
 	{
